# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "XTasks-API",
    "description": "Documentation for the API endpoints.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/upload-avatar": {
      "post": {
        "summary": "Upload user avatar",
        "operationId": "uploadAvatar",
        "tags": ["Users"],
        "responses": {
          "201": { "description": "Avatar uploaded successfully" },
          "401": {
            "description": "Only authenticated users can change avatar"
          },
          "500": { "description": "Error uploading avatar" }
        }
      }
    }
  }
}